requirements.txt
Flask
google-cloud-storage
google-genai
gunicorn
